requirements.txt
matplotlib==3.6.3
numpy==1.24.1
Pillow==9.4.0
pytorch_lightning==1.9.0
torch==1.13.1
torchio==0.18.86
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.9
